services/auth_service/pyproject.toml
[project]
name = "auth_service"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.122.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.1",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auth_service"]
